api/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==14.1
uvloop==0.21.0
httptools==0.6.4

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
asyncpg==0.30.0
alembic==1.17.1

# Authentication
python-jose==3.3.0
google-auth==2.32.0
PyJWT==2.10.1

# Utilities
pydantic==2.10.0
python-dotenv==1.0.1
httpx==0.27.2
requests
apscheduler==3.11.1

# Testing
pytest
pytest-asyncio

# Performance monitoring (optional)
# prometheus-client==0.21.0
